requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
nulltype==2.3.1
numpy==2.2.5
packaging==25.0
pandas==2.2.3
plaid-python==31.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
